# openapi.yaml
openapi: 3.0.3
info:
  title: PaySuite
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://paysuite.tech'
tags:
  -
    name: 'Payment Requests'
    description: "\nAPIs for managing payment requests"
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
  /api/v1/payments:
    post:
      summary: 'Create Payment Request'
      operationId: createPaymentRequest
      description: 'Creates a new payment request in the system.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  data:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    amount: 100.5
                    reference: INV2024001
                    status: pending
                    checkout_url: 'https://paysuite.test/checkout/550e8400-e29b-41d4-a716-446655440000'
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 550e8400-e29b-41d4-a716-446655440000
                      amount:
                        type: number
                        example: 100.5
                      reference:
                        type: string
                        example: INV2024001
                      status:
                        type: string
                        example: pending
                      checkout_url:
                        type: string
                        example: 'https://paysuite.test/checkout/550e8400-e29b-41d4-a716-446655440000'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Invalid input'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Invalid input'
      tags:
        - 'Payment Requests'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: numeric
                  description: 'Payment amount in MZN.'
                  example: '100.50'
                  nullable: false
                method:
                  type: string
                  description: 'optional Payment method.'
                  example: null
                  nullable: false
                  enum:
                    - credit_card
                    - mpesa
                    - emola
                reference:
                  type: string
                  description: 'Unique payment reference (max: 50 characters).'
                  example: INV2024001
                  nullable: false
                description:
                  type: string
                  description: 'optional Payment description (max: 255 characters).'
                  example: 'Payment for invoice INV2024001'
                  nullable: true
                return_url:
                  type: string
                  description: 'optional URL to redirect after payment completion.'
                  example: 'https://example.com/success'
                  nullable: true
                callback_url:
                  type: string
                  description: 'Must be a valid URL. Must not be greater than 255 characters.'
                  example: 'http://nolan.net/velit-aut-maiores-odit-laborum-repellat-totam-repellendus-et.html'
                  nullable: true
              required:
                - amount
                - reference
  '/api/v1/payments/{uuid}':
    get:
      summary: 'Get Payment Request'
      operationId: getPaymentRequest
      description: 'Returns the details of a specific payment request.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  data:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    amount: 100.5
                    reference: INV2024001
                    status: paid
                    transaction:
                      id: 1
                      status: completed
                      transaction_id: MPESA123456
                      paid_at: '2024-02-10T10:15:00.000000Z'
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 550e8400-e29b-41d4-a716-446655440000
                      amount:
                        type: number
                        example: 100.5
                      reference:
                        type: string
                        example: INV2024001
                      status:
                        type: string
                        example: paid
                      transaction:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          status:
                            type: string
                            example: completed
                          transaction_id:
                            type: string
                            example: MPESA123456
                          paid_at:
                            type: string
                            example: '2024-02-10T10:15:00.000000Z'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Payment request not found.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Payment request not found.'
      tags:
        - 'Payment Requests'
    parameters:
      -
        in: path
        name: uuid
        description: 'UUID of the payment request.'
        example: 550e8400-e29b-41d4-a716-446655440000
        required: true
        schema:
          type: string